requirements.txt
graphviz==0.19.2
langchain==0.1.5
langchain_core==0.1.18
langchain_openai==0.0.5
python-dotenv==1.0.1
